requirements.txt
Flask==3.0.0
elevenlabs==0.2.27
werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
